+++ x/gov/keeper/keeper_test.go
@@ -20,9 +20,9 @@ func TestKeeper_SaveGetPermissionsForRole(t *testing.T) {
 		nil, []types.PermValue{types.PermClaimValidator},
 	)
 
-	app.CustomGovKeeper.SetPermissionsForRole(ctx, types.RoleCouncilor, perm)
+	app.CustomGovKeeper.SetPermissionsForRole(ctx, types.RoleSudo, perm)
 
-	savedPerms := app.CustomGovKeeper.GetPermissionsForRole(ctx, types.RoleCouncilor)
+	savedPerms := app.CustomGovKeeper.GetPermissionsForRole(ctx, types.RoleSudo)
 	require.Equal(t, perm, savedPerms)
 }
 
@@ -88,4 +88,15 @@ func TestKeeper_AddPermissionToNetworkActor(t *testing.T) {
 	savedNetworkActor, err = app.CustomGovKeeper.GetNetworkActorByAddress(ctx, addr)
 	require.NoError(t, err)
 	require.True(t, savedNetworkActor.Permissions.IsWhitelisted(types.PermSetPermissions))
+}
+
+func TestKeeper_HasGenesisDefaultRoles(t *testing.T) {
+	app := simapp.Setup(false)
+	ctx := app.NewContext(false, tmproto.Header{})
+
+	roleSudo := app.CustomGovKeeper.GetPermissionsForRole(ctx, types.RoleSudo)
+	require.True(t, roleSudo.IsWhitelisted(types.PermSetPermissions))
+
+	roleValidator := app.CustomGovKeeper.GetPermissionsForRole(ctx, types.RoleValidator)
+	require.True(t, roleValidator.IsWhitelisted(types.PermClaimValidator))
 }