+++ x/gov/types/types_test.go
@@ -13,27 +13,28 @@ func TestPermissions_IsBlacklisted(t *testing.T) {
 	)
 
 	require.True(t, perms.IsBlacklisted(PermClaimValidator))
-	require.False(t, perms.IsBlacklisted(PermClaimGovernanceSeat))
+	require.False(t, perms.IsBlacklisted(PermSetPermissions))
 }
 
 func TestPermissions_IsWhitelisted(t *testing.T) {
-	perms := NewPermissions([]PermValue{PermClaimGovernanceSeat}, nil)
+	perms := NewPermissions([]PermValue{PermClaimValidator}, nil)
 
-	require.True(t, perms.IsWhitelisted(PermClaimGovernanceSeat))
-	require.False(t, perms.IsWhitelisted(PermClaimValidator))
+	require.True(t, perms.IsWhitelisted(PermClaimValidator))
+	require.False(t, perms.IsWhitelisted(PermSetPermissions))
 }
 
 func TestPermissions_AddWhitelist(t *testing.T) {
 	perms := NewPermissions(nil, nil)
 
-	require.False(t, perms.IsWhitelisted(PermClaimGovernanceSeat))
+	require.False(t, perms.IsWhitelisted(PermClaimValidator))
 
-	err := perms.AddToWhitelist(PermClaimGovernanceSeat)
+	err := perms.AddToWhitelist(PermSetPermissions)
 	require.NoError(t, err)
-	require.True(t, perms.IsWhitelisted(PermClaimGovernanceSeat))
+	require.True(t, perms.IsWhitelisted(PermSetPermissions))
 
 	// Add to whitelist value blacklisted gives error
-	perms.AddToBlacklist(PermClaimValidator)
+	err = perms.AddToBlacklist(PermClaimValidator)
+	require.NoError(t, err)
 
 	err = perms.AddToWhitelist(PermClaimValidator)
 	require.EqualError(t, err, "permission is already blacklisted")
@@ -42,10 +43,10 @@ func TestPermissions_AddWhitelist(t *testing.T) {
 func TestPermissions_AddBlacklist(t *testing.T) {
 	perms := NewPermissions(nil, nil)
 
-	require.False(t, perms.IsBlacklisted(PermClaimGovernanceSeat))
-	err := perms.AddToBlacklist(PermClaimGovernanceSeat)
+	require.False(t, perms.IsBlacklisted(PermSetPermissions))
+	err := perms.AddToBlacklist(PermSetPermissions)
 	require.NoError(t, err)
-	require.True(t, perms.IsBlacklisted(PermClaimGovernanceSeat))
+	require.True(t, perms.IsBlacklisted(PermSetPermissions))
 
 	// Add to blacklist when is whitelisted gives error
 	err = perms.AddToWhitelist(PermClaimValidator)